requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.7.4.post0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==3.0.1
attrs==24.3.0
audioop-lts==0.2.1
bcrypt==4.3.0
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
cryptography==44.0.3
decorator==5.1.1
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
Django==5.2.5
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imutils==0.5.4
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mss==10.0.0
multidict==6.1.0
numpy==2.3.2
openai==1.77.0
opencv-python==4.10.0.84
paramiko==3.5.1
pickle-mixin==1.0.2
prettytable==3.16.0
propcache==0.2.1
psutil==6.1.0
pycaw==20240210
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.1
PyNaCl==1.5.0
python-dotenv==1.1.0
python-multipart==0.0.20
pywin32==308
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.18.3
